[ simplification ] Get rid of names substitution
General name substitution on any `TTImp` (that can be in the user's data
types) is error prone and complex. Thus, it's a good idea to avoid it.

For this, we make the canonical signature type producer to work only
with unnamed signatures, thus providing names of the generator functions
to be borrowed from the target type data definition.